# backend/docs/swagger.yaml
definitions:
  types.Coordinates:
    properties:
      x:
        type: number
      "y":
        type: number
    required:
    - x
    - "y"
    type: object
  types.JSONResponse:
    properties:
      message:
        type: string
      statusCode:
        type: integer
    type: object
  types.Journey:
    properties:
      departureStationId:
        type: integer
      departureStationName:
        type: string
      departureTime:
        type: string
      distance:
        type: integer
      duration:
        type: integer
      id:
        type: integer
      returnStationId:
        type: integer
      returnStationName:
        type: string
      returnTime:
        type: string
    type: object
  types.Station:
    properties:
      address:
        type: string
      capacity:
        type: integer
      city:
        type: string
      coordinates:
        $ref: '#/definitions/types.Coordinates'
      departures:
        type: integer
      fid:
        type: integer
      id:
        type: integer
      name:
        type: string
      operator:
        type: string
      returns:
        type: integer
    type: object
  types.StationRequest:
    properties:
      address:
        type: string
      capacity:
        type: integer
      city:
        type: string
      coordinates:
        $ref: '#/definitions/types.Coordinates'
      id:
        type: integer
      name:
        type: string
      operator:
        type: string
    required:
    - address
    - city
    - coordinates
    - id
    - name
    type: object
host: localhost:8080
info:
  contact:
    name: Kimi Porthan
  description: Solita Dev Academy pre-assingment.
  title: Citybike 2023
  version: "0.1"
paths:
  /api/journeys/page/{page}:
    get:
      description: GET /api/journeys/{page} returns 10 * page amount of journeys.
      parameters:
      - description: Page number
        in: path
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Journey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.JSONResponse'
      summary: Get journeys by page.
      tags:
      - Journeys
  /api/stations:
    post:
      description: POST api/stations to add new station.
      parameters:
      - description: New station
        in: body
        name: addStationRequest
        required: true
        schema:
          $ref: '#/definitions/types.StationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Station'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.JSONResponse'
      summary: Add new station.
      tags:
      - Stations
  /api/stations/page/{page}:
    get:
      description: GET /api/stations/{page} returns 10 * page amount of stations.
      parameters:
      - description: Page number
        in: path
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Station'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.JSONResponse'
      summary: Get stations by page.
      tags:
      - Stations
  /api/stations/page/{page}/{name}:
    get:
      description: GET api/stations/{page}/{string}
      parameters:
      - description: Page number
        in: path
        name: page
        required: true
        type: integer
      - description: substring/name of station
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Station'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.JSONResponse'
      summary: Get stations by name.
      tags:
      - Stations
swagger: "2.0"